Started implemented the set command. history command implemented.
git-svn-id: svn://127.0.0.1/pplsvn@263 8e462c0f-ee70-4953-a96e-5db927b55778

+++ src/ppl_readconfig.c
@@ -140,7 +140,7 @@ void ReadConfigFile(char *ConfigFname)
         if ((fl=GetFloat(setvalue, &i), i==strlen(setvalue)))               settings_term_default .dpi           = fl;
         else {sprintf(temp_err_string, "Error in line %d of configuration file %s:\nIllegal value for setting DPI."          , linecounter, ConfigFname); ppl_warning(temp_err_string); continue; }
       else if (strcmp(setkey, "FONTSIZE"     )==0)
-        if ((fl=GetFloat(setvalue, &i), i==strlen(setvalue)))               settings_graph_default.FontSize      = (int)fl;
+        if ((fl=GetFloat(setvalue, &i), i==strlen(setvalue)))               settings_graph_default.FontSize      = fl;
         else {sprintf(temp_err_string, "Error in line %d of configuration file %s:\nIllegal value for setting FONTSIZE."     , linecounter, ConfigFname); ppl_warning(temp_err_string); continue; }
       else if (strcmp(setkey, "FUNCSTYLE"    )==0)
         if ((i=FetchSettingByName(setvalue,SW_STYLE_INT, SW_STYLE_STR ))>0) settings_graph_default.FuncStyle.style = i;
